refactor(context-engine): extract deferred-maintenance abort-signal wiring
Move the process-scoped SIGINT/SIGTERM abort-signal fan-out out of
context-engine-maintenance.ts into a self-contained sibling module so the
file stays within the max-lines budget. Pure code move plus import wiring;
preemption, write-fence, serialized maintenance, and per-task timeout
behavior are unchanged.
