+++ src/agents/embedded-agent-runner/context-engine-maintenance.ts
@@ -41,6 +41,12 @@ import { findActiveSessionTask } from "../session-async-task-status.js";
 import { SessionManager } from "../sessions/index.js";
 import { resolveContextEngineCapabilities } from "./context-engine-capabilities.js";
 import {
+  createDeferredTurnMaintenanceAbortSignal,
+  DEFERRED_TURN_MAINTENANCE_ABORT_STATE_KEY,
+  type DeferredTurnMaintenanceProcessLike,
+  unregisterDeferredTurnMaintenanceAbortSignalHandlers,
+} from "./context-engine-maintenance-abort-signal.js";
+import {
   createDeferredMaintenanceWriteFence,
   type DeferredMaintenanceWriteFence,
 } from "./context-engine-maintenance-fence.js";
@@ -54,9 +60,6 @@ const TURN_MAINTENANCE_LONG_WAIT_MS = 10_000;
 // A waiting user turn should only pay for cooperative abort cleanup, not the
 // background maintenance operation's otherwise unbounded runtime.
 const TURN_MAINTENANCE_PREEMPT_GRACE_MS = 1_000;
-const DEFERRED_TURN_MAINTENANCE_ABORT_STATE_KEY = Symbol.for(
-  "openclaw.contextEngineTurnMaintenanceAbortState",
-);
 type SessionManagerRewriteLock = <T>(operation: () => Promise<T> | T) => Promise<T>;
 
 type ContextEngineMaintenanceParams = {
@@ -88,26 +91,6 @@ type DeferredTurnMaintenanceScheduleParams = ContextEngineMaintenanceParams & {
 
 const activeDeferredTurnMaintenanceRuns = new Map<string, DeferredTurnMaintenanceRun>();
 
-type DeferredTurnMaintenanceSignal = "SIGINT" | "SIGTERM";
-type DeferredTurnMaintenanceProcessLike = Pick<NodeJS.Process, "on" | "off"> &
-  Partial<Pick<NodeJS.Process, "listenerCount" | "kill" | "pid">> & {
-    [DEFERRED_TURN_MAINTENANCE_ABORT_STATE_KEY]?: DeferredTurnMaintenanceAbortState;
-  };
-type DeferredTurnMaintenanceAbortState = {
-  controllers: Set<AbortController>;
-  cleanupHandlers: Map<DeferredTurnMaintenanceSignal, () => void>;
-};
-
-function unregisterDeferredTurnMaintenanceAbortSignalHandlers(
-  processLike: DeferredTurnMaintenanceProcessLike,
-  state: DeferredTurnMaintenanceAbortState,
-): void {
-  for (const [signal, handler] of state.cleanupHandlers) {
-    processLike.off(signal, handler);
-  }
-  state.cleanupHandlers.clear();
-}
-
 async function disposeDeferredMaintenanceContextEngine(
   contextEngine: ContextEngine,
 ): Promise<void> {
@@ -118,63 +101,6 @@ async function disposeDeferredMaintenanceContextEngine(
       errorMessage: formatErrorMessage(err),
     });
   }
-}
-
-function createDeferredTurnMaintenanceAbortSignal(params?: {
-  processLike?: DeferredTurnMaintenanceProcessLike;
-}): {
-  abortSignal: AbortSignal;
-  abort: (reason: Error) => void;
-  dispose: () => void;
-} {
-  const processLike = (params?.processLike ?? process) as DeferredTurnMaintenanceProcessLike;
-  const state = (processLike[DEFERRED_TURN_MAINTENANCE_ABORT_STATE_KEY] ??= {
-    controllers: new Set<AbortController>(),
-    cleanupHandlers: new Map<DeferredTurnMaintenanceSignal, () => void>(),
-  });
-  const handleTerminationSignal = (signalName: DeferredTurnMaintenanceSignal) => {
-    const shouldReraise = processLike.listenerCount?.(signalName) === 1;
-    for (const activeController of state.controllers) {
-      if (!activeController.signal.aborted) {
-        activeController.abort(
-          new Error(`received ${signalName} while waiting for deferred maintenance`),
-        );
-      }
-    }
-    state.controllers.clear();
-    unregisterDeferredTurnMaintenanceAbortSignalHandlers(processLike, state);
-    if (shouldReraise && typeof processLike.kill === "function") {
-      try {
-        processLike.kill(processLike.pid ?? process.pid, signalName);
-      } catch {
-        // Ignore shutdown-path failures.
-      }
-    }
-  };
-  if (state.cleanupHandlers.size === 0) {
-    for (const signal of ["SIGINT", "SIGTERM"] as const) {
-      const handler = () => handleTerminationSignal(signal);
-      state.cleanupHandlers.set(signal, handler);
-      processLike.on(signal, handler);
-    }
-  }
-
-  const controller = new AbortController();
-  state.controllers.add(controller);
-  return {
-    abortSignal: controller.signal,
-    abort: (reason) => {
-      if (!controller.signal.aborted) {
-        controller.abort(reason);
-      }
-    },
-    dispose: () => {
-      state.controllers.delete(controller);
-      if (state.controllers.size === 0) {
-        unregisterDeferredTurnMaintenanceAbortSignalHandlers(processLike, state);
-      }
-    },
-  };
 }
 
 function resetDeferredTurnMaintenanceStateForTest(): void {
